chore: move to cli
