+++ src/libecalc/domain/emission/time_series_intensity.py
@@ -1,7 +1,6 @@
 import math
 from typing import Any
 
-import pandas as pd
 from pydantic import BaseModel, field_validator
 from pydantic_core.core_schema import ValidationInfo
 
@@ -22,9 +21,3 @@ class TimeSeriesIntensity(BaseModel):
             # convert None to nan
             return [i if i is not None else math.nan for i in v]  # type: ignore[misc]
         return v
-
-    def to_dataframe(self, prefix: str = ""):
-        col_name = prefix if prefix else "value"
-        df = pd.DataFrame({col_name: self.values}, index=[p.start for p in self.periods])
-        df.index.name = "period"
-        return df
